requirements.txt
Flask==2.3.0
Flask-JWT-Extended==4.5.3
Flask-SQLAlchemy==3.0.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.0
python-dotenv==1.0.0
cryptography==41.0.0
Flask-Limiter==2.6.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9

email-validator==2.2.0